pyproject.toml
[tool.poetry]
name = "gpt-fine-tuning"
version = "0.1.0"
description = ""
authors = ["Sean <91415054+sean85120@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^0.28.0"
python-dotenv = "^1.0.0"
tiktoken = "^0.5.1"
numpy = "^1.25.2"
jupyter = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
